server/requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
black==23.11.0
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.3.0
click==8.2.1
click-plugins==1.1.1.2
cligj==0.7.2
colorama==0.4.6
colorlog==6.7.0
contourpy==1.3.2
cycler==0.12.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.104.1
filelock==3.19.1
fiona==1.10.1
flake8==6.1.0
Flask==2.3.3
Flask-Cors==4.0.0
fonttools==4.59.1
fsspec==2025.7.0
geopandas==1.1.1
h11==0.16.0
httptools==0.6.4
idna==3.4
ifaddr==0.2.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.2
joblib==1.5.1
jsonschema==4.19.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
loguru==0.7.2
MarkupSafe==3.0.2
matplotlib==3.10.5
mccabe==0.7.0
mega.py==1.0.8
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.4.2
numpy==2.2.6
PyYAML==6.0.1
opencv-python==4.12.0.88
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
pefile==2023.2.7
Pillow==10.1.0
platformdirs==4.3.8
pluggy==1.6.0
psutil==5.9.6
py-cpuinfo==9.0.0
pycodestyle==2.11.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.5.0
pydantic_core==2.14.1
pyflakes==3.1.0
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2025.3
pyodbc==4.0.39
pyogrio==0.11.1
pyparsing==3.2.3
PyPDF2==3.0.1
pyproj==3.7.1
PySocks==1.7.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2025.2
pywin32==306
pywin32-ctypes==0.2.3
PyYAML==6.0.1
referencing==0.36.2
requests==2.31.0
rpds-py==0.27.0
schedule==1.2.2
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
selenium==4.35.0
shapely==2.1.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
starlette==0.27.0
sympy==1.14.0
tabulate==0.9.0
tenacity==5.1.5
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tkcalendar==1.6.1
tomli==2.2.1
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.1.4
urllib3==2.5.0
uvicorn==0.24.0
watchfiles==1.1.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==2.3.7
win32_setctime==1.2.0
wsproto==1.2.0
XlsxWriter==3.1.9
zeroconf==0.131.0